+++ Notes/Change Log.md
@@ -1,5 +1,6 @@
 #### 19.08.25
 - A A-Star Grid map to help with creating the Unity grid
+- A-Star grid created in Unity
 ---
 #### 18.08.25
 - Fixed pursue mover for this style of enemy movement